% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra's Algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A data frame representing the weighted graph with columns:
- v1: The starting node of an edge.
- v2: The ending node of an edge.
- w: The weight of the edge.}

\item{init_node}{The initial node from which to find the shortest paths.}
}
\value{
A numeric vector of length equal to the number of nodes in the graph.
 Each element represents the shortest distance from the initial node to the
 corresponding node in the graph.
}
\description{
The algorithm takes a graph and an initial node and calculates the shortest path from the
}
\references{
Wikipedia: Dijkstra's Algorithm - https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm
}
