% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean Algorithm for Greatest Common Divisor (GCD)}
\usage{
euclidean(n, m)
}
\arguments{
\item{n}{A positive numeric value.}

\item{m}{A positive numeric value.}
}
\value{
The greatest common divisor (GCD) of n and m, which is a positive integer.
}
\description{
This function calculates the greatest common divisor (GCD) of two positive integers
using the Euclidean algorithm.
}
\references{
Wikipedia: Euclidean Algorithm - https://en.wikipedia.org/wiki/Euclidean_algorithm
}
