% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{wiki graph data}
\format{
An object of class \code{data.frame} with 18 rows and 3 columns.
}
\usage{
wiki_graph
}
\description{
wiki graph data
}
\references{
\url{https://en.wikipedia.org/wiki/Graph}
}
\keyword{data}
